just check _parsed.json
